third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-008-expected.txt


FAIL CSS Test: @viewport constrained - min-zoom greater than max-zoom assert_equals: expected 400 but got 800
